+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * [Fix] Fixes `%sqlcmd plot` when `--table` or `--column` have spaces (#409)
 * [Feature] Add sticky first column styling to sqlcmd profile command
 * [Feature] Improve boxplot performance [#152]
+* [Feature] Adds `--persist-replace` argument
+
 ## 0.7.1 (2023-04-19)
 
 * [Feature] Upgrades SQLAlchemy version to 2